+++ src/views/base-view.js
@@ -1,6 +1,5 @@
 'use strict';
 
-var assign = require('../lib/assign').assign;
 var classList = require('@braintree/class-list');
 var DropinError = require('../lib/dropin-error');
 var errors = require('../constants').errors;
@@ -8,7 +7,7 @@ var errors = require('../constants').errors;
 function BaseView(options) {
   options = options || {};
 
-  assign(this, options);
+  Object.assign(this, options);
 }
 
 BaseView.prototype.getElementById = function (id) {