+++ src/views/payment-sheet-views/paypal-credit-view.js
@@ -1,6 +1,5 @@
 'use strict';
 
-var assign = require('../../lib/assign').assign;
 var paymentOptionIDs = require('../../constants').paymentOptionIDs;
 var BasePayPalView = require('./base-paypal-view');
 
@@ -19,7 +18,7 @@ PayPalCreditView.isEnabled = function (options) {
     return Promise.resolve(false);
   }
 
-  return BasePayPalView.isEnabled(assign({
+  return BasePayPalView.isEnabled(Object.assign({
     viewID: PayPalCreditView.ID
   }, options));
 };