+++ test/integration/update-configuration.test.js
@@ -62,12 +62,6 @@ describe('Drop-in#updateConfiguration', function () {
   });
 
   it('removes authorized PayPal account when configuration is updated @paypal', function () {
-    if (browser.name() === 'INTERNET EXPLORER') {
-      this.skip('Fails for unknown reasons on IE');
-
-      return;
-    }
-
     browser.start({
       paypal: 'default',
       showUpdatePayPalMenu: true