+++ wdio.conf.js
@@ -72,26 +72,8 @@ let capabilities = [
   //    }
   //  }
   //},
-  {
-    ...desktopCapabilities,
-    browser: 'IE',
-    browserName: 'IE 11',
-    browser_version: '11.0',
-    'browserstack.selenium_version' : '3.141.5',
-    // https://stackoverflow.com/a/42340325/7851516
-    'browserstack.bfcache': '0',
-    // don't update this! There's a weird bug in the
-    // 64 bit ie driver that prevents the shift key
-    // from working which means that an email can
-    // never be entered because the "@" key cannot
-    // be entered. This doesn't occur in the 32 bit
-    // version, so we pin to that
-    'browserstack.ie.arch' : 'x32'
-  },
-  // TODO edge has been pretty flaky
-  // disable it for now to make the transition
-  // to browserstack, but look into re-enabling
-  // it once the transition is complete
+  // TODO edge has been pretty flaky—disable it for now to make the transition
+  //  to browserstack, but look into re-enabling it once the transition is complete
   //{
   //  ...desktopCapabilities,
   //  browserName: 'Microsoft Edge',