+++ CHANGELOG.md
@@ -6,7 +6,7 @@ CHANGELOG
 * PayPal can now override global autovaulting preferences with `paypal.vault.autoVault`
 
 _Breaking Changes_
-* Drop support for IE 9-10
+* Drop support for Internet Explorer
 * Whenever a new payment method is added, all other unvaulted payment methods will be removed from available payment methods
 * Payment methods will no longer automatically be hidden or error on setup if merchant is not enabled for them
 * Credit card view changes